+++ passerelle_imio_ia_aes/models.py
@@ -51,15 +51,13 @@ import random
 import requests
 try:
     from urllib.parse import urlparse
+    from urllib.parse import urljoin
 except ImportError:
     import urlparse
-try:
-    import xmlrpc.client
-    from xmlrpc.client import ServerProxy
-except ImportError:
-    import xmlrpclib
-    # noinspection PyCompatibility
-    from xmlrpclib import ServerProxy
+
+import xmlrpc.client
+from xmlrpc.client import ServerProxy
+
 
 from datetime import datetime as dt
 from decimal import Decimal
@@ -67,6 +65,7 @@ from django.conf import settings
 from django.db import models
 from django.http import HttpResponse
 from django.utils import timezone
+from django.utils.encoding import force_text
 from django.utils.translation import ugettext_lazy as _
 from passerelle.base.models import BaseResource
 from passerelle.base.signature import sign_url
@@ -74,7 +73,7 @@ from passerelle.utils.api import endpoint
 
 
 def format_type(t):
-    return {"id": unicode(t), "text": unicode(t)}
+    return {"id": force_text(t), "text": force_text(t)}
 
 
 def format_file(f):
@@ -94,7 +93,7 @@ class FileNotFoundError(Exception):
 # http://local-formulaires.example.net/travaux/demo-cb-aes/1/jump/trigger/validate
 
 
-class ProxiedTransport(xmlrpclib.Transport):
+class ProxiedTransport(xmlrpc.client.Transport):
     def set_proxy(self, proxy):
         self.proxy = proxy
 
@@ -167,32 +166,14 @@ class IImioIaAes(BaseResource):
         description="Tester la connexion avec AES",
     )
     def tst_connexion(self, request):
-        test = None
-        try:
-            test = self.get_aes_server().execute_kw(
-                self.database_name,
-                self.get_aes_user_id(),
-                self.password,
-                "aes_api.aes_api",
-                "hello_world",
-                [],
-            )
-        except Exception:
-            p = ProxiedTransport()
-            p.set_proxy("10.9.200.215:9069")
-            # server = xmlrpclib.ServerProxy('http://time.xmlrpc.com/RPC2', transport=p)
-            server = ServerProxy(
-                "{}/xmlrpc/2/object".format(self.server_url), transport=p
-            )
-            test = server.execute_kw(
-                self.database_name,
-                self.get_aes_user_id(),
-                self.password,
-                "aes_api.aes_api",
-                "hello_world",
-                [],
-            )
-            # print server.currentTime.getCurrentTime()
+        test = self.get_aes_server().execute_kw(
+               self.database_name,
+               self.get_aes_user_id(),
+               self.password,
+               "aes_api.aes_api",
+               "hello_world",
+               [],
+        )
         return test
 
     @endpoint(
@@ -247,7 +228,7 @@ class IImioIaAes(BaseResource):
     def post_child_meal(self, request, *args, **kwargs):
         # data = dict([(x, request.GET[x]) for x in request.GET.keys()])
         if request.body:
-            occurences_load = json.loads(request.body)
+            occurences_load = json.loads(request.body.decode("utf-8"))
         is_add = self.get_aes_server().execute_kw(
             self.database_name,
             self.get_aes_user_id(),
@@ -266,7 +247,7 @@ class IImioIaAes(BaseResource):
     )
     def post_child_health_sheet(self, request, *args, **kwargs):
         try:
-            occurences_load = json.loads(request.body)
+            occurences_load = json.loads(request.body.decode("utf-8"))
             fields = occurences_load.get("fields")
         except ValueError as e:
             raise ValueError(e.message)
@@ -372,6 +353,36 @@ class IImioIaAes(BaseResource):
         except Exception:
             return False
 
+
+    @endpoint(
+        serializer_type="json-api",
+        perm="can_access",
+        description="Récupérer les enfants pour le parent connecté, en interrogeant le RN du parentplutôt que son mail",
+        parameters={
+            "nrn": {
+                "description": "Numéro de registre national d'un parent AES/TS",
+                "example_value": "00000000097",
+            }
+        },
+    )
+    def get_children_by_parent_nrn(self, request, **kwargs):
+        parent = {"nrn": request.GET["nrn"]}
+        if parent["nrn"] == "":
+            return "No nrn"
+        try:
+            children = self.get_aes_server().execute_kw(
+                self.database_name,
+                self.get_aes_user_id(),
+                self.password,
+                "aes_api.aes_api",
+                "get_children",
+                [parent],
+            )
+            return children
+        except Exception:
+            return False
+
+
     @endpoint(
         serializer_type="json-api",
         perm="can_access",
@@ -406,7 +417,7 @@ class IImioIaAes(BaseResource):
     )
     def parent_registration(self, request, **kwargs):
         if request.body:
-            parent = json.loads(request.body)
+            parent = json.loads(request.body.decode("utf-8"))
         registration_id = self.get_aes_server().execute_kw(
             self.database_name,
             self.get_aes_user_id(),
@@ -424,7 +435,7 @@ class IImioIaAes(BaseResource):
     )
     def get_parent_id(self, request, email=None, nrn=None):
         if request.body:
-            params = json.loads(request.body)
+            params = json.loads(request.body.decode("utf-8"))
             data = {"email": params.get("email"), "nrn": params.get("nrn")}
         else:
             data = dict([(x, request.GET[x]) for x in request.GET.keys()])
@@ -446,9 +457,9 @@ class IImioIaAes(BaseResource):
     )
     def child_registration(self, request, **kwargs):
         if request.body:
-            params = json.loads(request.body)
+            params = json.loads(request.body.decode("utf-8"))
         parent_id = self.get_parent_id(request)
-        parentid = {u"parentid": unicode(parent_id.get("id"))}
+        parentid = {u"parentid": force_text(parent_id.get("id"))}
         params.update(parentid)
         params = {
             key: value
@@ -495,9 +506,9 @@ class IImioIaAes(BaseResource):
 
         idp_service = list(settings.KNOWN_SERVICES['authentic'].values())[0]
         api_url = sign_url(
-                urlparse.urljoin(
+                urljoin(
                     idp_service['url'],
-                    'api/users/?email=%s&orig=%s' % (urllib.quote_plus(request.GET["email"]), idp_service.get('orig'))),
+                    'api/users/?email=%s&orig=%s' % (urllib.parse.quote_plus(request.GET["email"]), idp_service.get('orig'))),
                 key=idp_service.get('secret'))
         r = self.requests.get(api_url)
         nrn = (
@@ -597,7 +608,7 @@ class IImioIaAes(BaseResource):
     def add_registration_child(self, request, *args, **kwargs):
         data = dict([(x, request.GET[x]) for x in request.GET.keys()])
         if request.body:
-            occurences_load = json.loads(request.body)
+            occurences_load = json.loads(request.body.decode("utf-8"))
         is_registration_child = self.get_aes_server().execute_kw(
             self.database_name,
             self.get_aes_user_id(),
@@ -625,7 +636,7 @@ class IImioIaAes(BaseResource):
     )
     def add_registration_child_plaine_output(self, request, *args, **kwargs):
         if request.body:
-            occurences_load = json.loads(request.body)
+            occurences_load = json.loads(request.body.decode("utf-8"))
         else:
             occurences_load = kwargs
         if not isinstance(occurences_load.get("data"), list):
@@ -701,7 +712,7 @@ class IImioIaAes(BaseResource):
         if debug is True:
             return {"data": [{"name": "Theme Label"}]}
         if getattr(request, "body", None) is not None:
-            params = json.loads(request.body)
+            params = json.loads(request.body.decode("utf-8"))
             activity_id = params.get("activity_id")
             week_number = params.get("week_number")
         data = {"activity_id": activity_id, "week_number": week_number}
@@ -864,7 +875,7 @@ class IImioIaAes(BaseResource):
     )
     def validate_form(self, request):
         if request.body:
-            data = json.loads(request.body)
+            data = json.loads(request.body.decode("utf-8"))
         self.get_aes_server().execute_kw(
             self.database_name,
             self.get_aes_user_id(),
@@ -883,7 +894,7 @@ class IImioIaAes(BaseResource):
     )
     def close_plaines_reservation(self, request):
         if request.body:
-            data = json.loads(request.body)
+            data = json.loads(request.body.decode("utf-8"))
         pay = self.get_aes_server().execute_kw(
             self.database_name,
             self.get_aes_user_id(),
@@ -902,7 +913,7 @@ class IImioIaAes(BaseResource):
     )
     def free_up_places(self, request):
         if request.body:
-            data = json.loads(request.body)
+            data = json.loads(request.body.decode("utf-8"))
         self.get_aes_server().execute_kw(
             self.database_name,
             self.get_aes_user_id(),
@@ -921,7 +932,7 @@ class IImioIaAes(BaseResource):
     )
     def pay_prepaid(self, request, amount=None, parent_id=None, form_id=None):
         if request.body:
-            params = json.loads(request.body)
+            params = json.loads(request.body.decode("utf-8"))
         else:
             params = dict([(x, request.GET[x]) for x in request.GET.keys()])
         data = {
@@ -1036,7 +1047,7 @@ class IImioIaAes(BaseResource):
         parameters={
             "invoice_id": {
                 "description": _("Invoice identifier"),
-                "example_value": invoices.keys()[0],
+                "example_value": list(invoices)[0],
             }
         },
     )
@@ -1051,7 +1062,7 @@ class IImioIaAes(BaseResource):
         parameters={
             "invoice_id": {
                 "description": _("Invoice identifier"),
-                "example_value": invoices.keys()[0],
+                "example_value": list(invoices)[0],
             }
         },
     )
@@ -1094,12 +1105,12 @@ class IImioIaAes(BaseResource):
         parameters={
             "invoice_id": {
                 "description": _("Invoice identifier"),
-                "example_value": invoices.keys()[0],
+                "example_value": list(invoices)[0],
             }
         },
     )
     def invoice_pay(self, request, invoice_id, NameID=None, **kwargs):
-        response = json.loads(request.body)
+        response = json.loads(request.body.decode("utf-8"))
         # ast.literal_eval(request.body)
         response["id"] = invoice_id
         aes_resp = self.get_aes_server().execute_kw(